Generalize feature activation and powerlevel code
