+++ app/lib/features/space/settings/pages/apps_settings_page.dart
@@ -1,13 +1,12 @@
 import 'package:acter/common/providers/room_providers.dart';
 import 'package:acter/common/providers/space_providers.dart';
-import 'package:acter/common/toolkit/buttons/primary_action_button.dart';
 import 'package:acter/common/utils/utils.dart';
 import 'package:acter/common/widgets/with_sidebar.dart';
+import 'package:acter/features/space/actions/set_acter_feature.dart';
+import 'package:acter/features/space/actions/update_feature_power_level.dart';
 import 'package:acter/features/space/settings/widgets/space_settings_menu.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:logging/logging.dart';
@@ -100,7 +99,7 @@ class SpaceAppsSettingsPage extends ConsumerWidget {
                 : 'Custom ($currentPw)';
             moreSections.add(
               SettingsSection(
-                title: const Text('Updates'),
+                title: Text(L10n.of(context).updates),
                 tiles: [
                   SettingsTile(
                     enabled: canEdit,
@@ -110,12 +109,15 @@ class SpaceAppsSettingsPage extends ConsumerWidget {
                     ),
                     trailing:
                         currentPw != null ? Text(pwText) : Text(defaultDesc),
-                    onPressed: (context) async => await onNewsLevelChange(
+                    onPressed: (context) async =>
+                        await updateFeatureLevelChange(
                       context,
                       maxPowerLevel,
                       currentPw,
                       space,
                       powerLevels,
+                      powerLevels.newsKey(),
+                      L10n.of(context).updates,
                     ),
                   ),
                   SettingsTile.switchTile(
@@ -146,12 +148,15 @@ class SpaceAppsSettingsPage extends ConsumerWidget {
                     ),
                     trailing:
                         currentPw != null ? Text(pwText) : Text(defaultDesc),
-                    onPressed: (context) async => await onPinLevelChange(
+                    onPressed: (context) async =>
+                        await updateFeatureLevelChange(
                       context,
                       maxPowerLevel,
                       currentPw,
                       space,
                       powerLevels,
+                      powerLevels.pinsKey(),
+                      L10n.of(context).pins,
                     ),
                   ),
                   SettingsTile.switchTile(
@@ -183,12 +188,14 @@ class SpaceAppsSettingsPage extends ConsumerWidget {
                     trailing:
                         currentPw != null ? Text(pwText) : Text(defaultDesc),
                     onPressed: (context) async =>
-                        await onCalendarEventLevelChange(
+                        await updateFeatureLevelChange(
                       context,
                       maxPowerLevel,
                       currentPw,
                       space,
                       powerLevels,
+                      powerLevels.eventsKey(),
+                      L10n.of(context).events,
                     ),
                   ),
                   SettingsTile(
@@ -233,12 +240,15 @@ class SpaceAppsSettingsPage extends ConsumerWidget {
                     trailing: taskListCurrentPw != null
                         ? Text(pwTextTL)
                         : Text(defaultDesc),
-                    onPressed: (context) async => await onTaskListLevelChange(
+                    onPressed: (context) async =>
+                        await updateFeatureLevelChange(
                       context,
                       maxPowerLevel,
                       taskListCurrentPw,
                       space,
                       powerLevels,
+                      powerLevels.taskListsKey(),
+                      L10n.of(context).taskList,
                     ),
                   ),
                   SettingsTile(
@@ -250,12 +260,15 @@ class SpaceAppsSettingsPage extends ConsumerWidget {
                     trailing: tasksCurrentPw != null
                         ? Text(pwTextT)
                         : Text(defaultDesc),
-                    onPressed: (context) async => await onTaskLevelChange(
+                    onPressed: (context) async =>
+                        await updateFeatureLevelChange(
                       context,
                       maxPowerLevel,
                       tasksCurrentPw,
                       space,
                       powerLevels,
+                      powerLevels.tasksKey(),
+                      L10n.of(context).tasks,
                     ),
                   ),
                   SettingsTile.switchTile(
@@ -285,12 +298,13 @@ class SpaceAppsSettingsPage extends ConsumerWidget {
                       enabled: canEdit,
                       description: const Text('Post space-wide updates'),
                       initialValue: news.active(),
-                      onToggle: (newVal) async => await onNewsToggle(
+                      onToggle: (newVal) async => await setActerFeature(
                         context,
-                        news,
                         newVal,
                         appSettings,
                         space,
+                        SpaceFeature.updates,
+                        L10n.of(context).updates,
                       ),
                     ),
                     SettingsTile.switchTile(
@@ -298,12 +312,13 @@ class SpaceAppsSettingsPage extends ConsumerWidget {
                       enabled: canEdit,
                       description: const Text('Pin important information'),
                       initialValue: pins.active(),
-                      onToggle: (newVal) async => onPinToggle(
+                      onToggle: (newVal) async => await setActerFeature(
                         context,
-                        pins,
                         newVal,
                         appSettings,
                         space,
+                        SpaceFeature.pins,
+                        L10n.of(context).pins,
                       ),
                     ),
                     SettingsTile.switchTile(
@@ -311,12 +326,13 @@ class SpaceAppsSettingsPage extends ConsumerWidget {
                       enabled: canEdit,
                       description: const Text('Calender with Events'),
                       initialValue: events.active(),
-                      onToggle: (newVal) async => await onCalendarEventToggle(
+                      onToggle: (newVal) async => await setActerFeature(
                         context,
-                        events,
                         newVal,
                         appSettings,
                         space,
+                        SpaceFeature.events,
+                        L10n.of(context).event,
                       ),
                     ),
                     SettingsTile.switchTile(
@@ -325,12 +341,13 @@ class SpaceAppsSettingsPage extends ConsumerWidget {
                       enabled: canEdit,
                       description: const Text('ToDo-Lists & Tasks'),
                       initialValue: tasks.active(),
-                      onToggle: (newVal) async => await onTaskToggle(
+                      onToggle: (newVal) async => await setActerFeature(
                         context,
-                        tasks,
                         newVal,
                         appSettings,
                         space,
+                        SpaceFeature.tasks,
+                        L10n.of(context).tasks,
                       ),
                     ),
                     ...labActions,
@@ -350,539 +367,5 @@ class SpaceAppsSettingsPage extends ConsumerWidget {
         },
       ),
     );
-  }
-
-  Future<void> onNewsLevelChange(
-    BuildContext context,
-    int maxPowerLevel,
-    int? currentPw,
-    Space space,
-    RoomPowerLevels powerLevels,
-  ) async {
-    final newPowerLevel = await showDialog<int?>(
-      context: context,
-      builder: (BuildContext context) => ChangePowerLevel(
-        featureName: 'Updates',
-        currentPowerLevelName:
-            maxPowerLevel == 100 ? powerLevelName(currentPw) : 'Custom',
-        currentPowerLevel: currentPw,
-      ),
-    );
-    if (newPowerLevel == currentPw) return;
-    if (!context.mounted) {
-      EasyLoading.dismiss();
-      return;
-    }
-    EasyLoading.show(status: L10n.of(context).changingPowerLevelOf('Updates'));
-    try {
-      await space.updateFeaturePowerLevels(
-        powerLevels.newsKey(),
-        newPowerLevel,
-      );
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showToast(L10n.of(context).powerLevelSubmitted('Updates'));
-    } catch (e, s) {
-      _log.severe('Failed to change power level of Updates', e, s);
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showError(
-        L10n.of(context).failedToChangePowerLevel(e),
-        duration: const Duration(seconds: 3),
-      );
-    }
-  }
-
-  Future<void> onNewsToggle(
-    BuildContext context,
-    NewsSettings news,
-    bool newVal,
-    ActerAppSettings appSettings,
-    Space space,
-  ) async {
-    EasyLoading.show(status: L10n.of(context).changingSettingOf('Updates'));
-    try {
-      final updated = news.updater();
-      updated.active(newVal);
-      final builder = appSettings.updateBuilder();
-      builder.news(updated.build());
-      await space.updateAppSettings(builder);
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showToast(L10n.of(context).changedSettingOf('Updates'));
-    } catch (e, s) {
-      _log.severe('Failed to change setting of Updates', e, s);
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showError(
-        L10n.of(context).failedToToggleSettingOf('Updates', e),
-        duration: const Duration(seconds: 3),
-      );
-    }
-  }
-
-  Future<void> onPinLevelChange(
-    BuildContext context,
-    int maxPowerLevel,
-    int? currentPw,
-    Space space,
-    RoomPowerLevels powerLevels,
-  ) async {
-    final newPowerLevel = await showDialog<int?>(
-      context: context,
-      builder: (BuildContext context) => ChangePowerLevel(
-        featureName: 'Pin',
-        currentPowerLevelName:
-            maxPowerLevel == 100 ? powerLevelName(currentPw) : 'Custom',
-        currentPowerLevel: currentPw,
-      ),
-    );
-    if (newPowerLevel == currentPw) return;
-    if (!context.mounted) {
-      EasyLoading.dismiss();
-      return;
-    }
-    EasyLoading.show(status: L10n.of(context).changingSettingOf('Pins'));
-    try {
-      await space.updateFeaturePowerLevels(
-        powerLevels.pinsKey(),
-        newPowerLevel,
-      );
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showToast(L10n.of(context).powerLevelSubmitted('Pins'));
-    } catch (e, s) {
-      _log.severe('Failed to change power level of Pins', e, s);
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showError(
-        L10n.of(context).failedToChangePowerLevel(e),
-        duration: const Duration(seconds: 3),
-      );
-    }
-  }
-
-  Future<void> onPinToggle(
-    BuildContext context,
-    PinsSettings pins,
-    bool newVal,
-    ActerAppSettings appSettings,
-    Space space,
-  ) async {
-    EasyLoading.show(status: L10n.of(context).changingSettingOf('Pins'));
-    try {
-      final updated = pins.updater();
-      updated.active(newVal);
-      final builder = appSettings.updateBuilder();
-      builder.pins(updated.build());
-      await space.updateAppSettings(builder);
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showToast(L10n.of(context).changedSettingOf('Pins'));
-    } catch (e, s) {
-      _log.severe('Failed to change setting of Pins', e, s);
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showError(
-        L10n.of(context).failedToToggleSettingOf('Pins', e),
-        duration: const Duration(seconds: 3),
-      );
-    }
-  }
-
-  Future<void> onCalendarEventLevelChange(
-    BuildContext context,
-    int maxPowerLevel,
-    int? currentPw,
-    Space space,
-    RoomPowerLevels powerLevels,
-  ) async {
-    final newPowerLevel = await showDialog<int?>(
-      context: context,
-      builder: (BuildContext context) => ChangePowerLevel(
-        featureName: 'Calendar Events',
-        currentPowerLevelName:
-            maxPowerLevel == 100 ? powerLevelName(currentPw) : 'Custom',
-        currentPowerLevel: currentPw,
-      ),
-    );
-    if (newPowerLevel == currentPw) return;
-    if (!context.mounted) {
-      EasyLoading.dismiss();
-      return;
-    }
-    EasyLoading.show(status: L10n.of(context).changingPowerLevelOf('Events'));
-    try {
-      await space.updateFeaturePowerLevels(
-        powerLevels.eventsKey(),
-        newPowerLevel,
-      );
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showToast(L10n.of(context).powerLevelSubmitted('Events'));
-    } catch (e, s) {
-      _log.severe('Failed to change power level of Events', e, s);
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showError(
-        L10n.of(context).failedToChangePowerLevel(e),
-        duration: const Duration(seconds: 3),
-      );
-    }
-  }
-
-  Future<void> onCalendarEventToggle(
-    BuildContext context,
-    EventsSettings events,
-    bool newVal,
-    ActerAppSettings appSettings,
-    Space space,
-  ) async {
-    EasyLoading.show(status: L10n.of(context).changingSettingOf('Events'));
-    try {
-      final updated = events.updater();
-      updated.active(newVal);
-      final builder = appSettings.updateBuilder();
-      builder.events(updated.build());
-      await space.updateAppSettings(builder);
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showToast(L10n.of(context).changedSettingOf('Events'));
-    } catch (e, s) {
-      _log.severe('Failed to change setting of Events', e, s);
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showError(
-        L10n.of(context).failedToToggleSettingOf('Events', e),
-        duration: const Duration(seconds: 3),
-      );
-    }
-  }
-
-  Future<void> onTaskListLevelChange(
-    BuildContext context,
-    int maxPowerLevel,
-    int? currentPw,
-    Space space,
-    RoomPowerLevels powerLevels,
-  ) async {
-    final newPowerLevel = await showDialog<int?>(
-      context: context,
-      builder: (BuildContext context) => ChangePowerLevel(
-        featureName: 'Task Lists',
-        currentPowerLevelName:
-            maxPowerLevel == 100 ? powerLevelName(currentPw) : 'Custom',
-        currentPowerLevel: currentPw,
-      ),
-    );
-    if (newPowerLevel == currentPw) return;
-    if (!context.mounted) {
-      EasyLoading.dismiss();
-      return;
-    }
-    EasyLoading.show(
-      status: L10n.of(context).changingPowerLevelOf('Tasklists'),
-    );
-    try {
-      await space.updateFeaturePowerLevels(
-        powerLevels.taskListsKey(),
-        newPowerLevel,
-      );
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showToast(L10n.of(context).powerLevelSubmitted('Tasklists'));
-    } catch (e, s) {
-      _log.severe('Failed to change power level of Tasklists', e, s);
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showError(
-        L10n.of(context).failedToChangePowerLevel(e),
-        duration: const Duration(seconds: 3),
-      );
-    }
-  }
-
-  Future<void> onTaskLevelChange(
-    BuildContext context,
-    int maxPowerLevel,
-    int? currentPw,
-    Space space,
-    RoomPowerLevels powerLevels,
-  ) async {
-    final newPowerLevel = await showDialog<int?>(
-      context: context,
-      builder: (BuildContext context) => ChangePowerLevel(
-        featureName: 'Tasks',
-        currentPowerLevelName:
-            maxPowerLevel == 100 ? powerLevelName(currentPw) : 'Custom',
-        currentPowerLevel: currentPw,
-      ),
-    );
-    if (newPowerLevel == currentPw) return;
-    if (!context.mounted) {
-      EasyLoading.dismiss();
-      return;
-    }
-    EasyLoading.show(status: L10n.of(context).changingPowerLevelOf('Updates'));
-    try {
-      await space.updateFeaturePowerLevels(
-        powerLevels.tasksKey(),
-        newPowerLevel,
-      );
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showToast(L10n.of(context).powerLevelSubmitted('Tasks'));
-    } catch (e, s) {
-      _log.severe('Failed to change power level of Tasks', e, s);
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showError(
-        L10n.of(context).failedToChangePowerLevel(e),
-        duration: const Duration(seconds: 3),
-      );
-    }
-  }
-
-  Future<void> onTaskToggle(
-    BuildContext context,
-    TasksSettings tasks,
-    bool newVal,
-    ActerAppSettings appSettings,
-    Space space,
-  ) async {
-    EasyLoading.show(status: L10n.of(context).changingSettingOf('Tasks'));
-    try {
-      final updated = tasks.updater();
-      updated.active(newVal);
-      final builder = appSettings.updateBuilder();
-      builder.tasks(updated.build());
-      await space.updateAppSettings(builder);
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showToast(L10n.of(context).changedSettingOf('Tasks'));
-    } catch (e, s) {
-      _log.severe('Failed to change setting of Tasks', e, s);
-      if (!context.mounted) {
-        EasyLoading.dismiss();
-        return;
-      }
-      EasyLoading.showError(
-        L10n.of(context).failedToToggleSettingOf('Tasks', e),
-        duration: const Duration(seconds: 3),
-      );
-    }
-  }
-}
-
-class ChangePowerLevel extends StatefulWidget {
-  final String featureName;
-  final int? currentPowerLevel;
-  final String currentPowerLevelName;
-
-  const ChangePowerLevel({
-    super.key,
-    required this.featureName,
-    required this.currentPowerLevelName,
-    this.currentPowerLevel,
-  });
-
-  @override
-  State<ChangePowerLevel> createState() => _ChangePowerLevelState();
-}
-
-class _ChangePowerLevelState extends State<ChangePowerLevel> {
-  final TextEditingController dropDownMenuCtrl = TextEditingController();
-  final GlobalKey<FormState> _formKey =
-      GlobalKey<FormState>(debugLabel: 'change power level form');
-
-  String? currentMemberStatus;
-  int? customValue;
-
-  @override
-  void initState() {
-    super.initState();
-    setState(() => currentMemberStatus = widget.currentPowerLevelName);
-  }
-
-  void _updateMembershipStatus(String? value) {
-    if (mounted) {
-      setState(() => currentMemberStatus = value);
-    }
-  }
-
-  void _newCustomLevel(String? value) {
-    if (mounted) {
-      setState(() {
-        if (value != null) {
-          customValue = int.tryParse(value);
-        } else {
-          customValue = null;
-        }
-      });
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final memberStatus = widget.currentPowerLevelName;
-    final currentPowerLevel = widget.currentPowerLevel;
-    return AlertDialog(
-      title: const Text('Update Power level'),
-      content: Form(
-        key: _formKey,
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            const Text('Change the power level of'),
-            Text(widget.featureName),
-            // Row(
-            //   children: [
-            currentPowerLevel != null
-                ? Text('from $memberStatus ($currentPowerLevel) to ')
-                : const Text('from default to'),
-            Padding(
-              padding: const EdgeInsets.all(5),
-              child: DropdownButtonFormField(
-                value: currentMemberStatus,
-                onChanged: _updateMembershipStatus,
-                items: const [
-                  DropdownMenuItem(
-                    value: 'Admin',
-                    child: Text('Admin'),
-                  ),
-                  DropdownMenuItem(
-                    value: 'Mod',
-                    child: Text('Moderator'),
-                  ),
-                  DropdownMenuItem(
-                    value: 'Regular',
-                    child: Text('Regular'),
-                  ),
-                  DropdownMenuItem(
-                    value: 'None',
-                    child: Text('None'),
-                  ),
-                  DropdownMenuItem(
-                    value: 'Custom',
-                    child: Text('Custom'),
-                  ),
-                ],
-              ),
-            ),
-            Visibility(
-              visible: currentMemberStatus == 'Custom',
-              child: Padding(
-                padding: const EdgeInsets.all(5),
-                child: TextFormField(
-                  decoration: const InputDecoration(
-                    hintText: 'any number',
-                    labelText: 'Custom power level',
-                  ),
-                  onChanged: _newCustomLevel,
-                  initialValue: currentPowerLevel.toString(),
-                  keyboardType:
-                      const TextInputType.numberWithOptions(signed: true),
-                  inputFormatters: <TextInputFormatter>[
-                    FilteringTextInputFormatter.digitsOnly,
-                  ], // Only numbers
-                  validator: (String? value) {
-                    return currentMemberStatus == 'Custom' &&
-                            (value == null || int.tryParse(value) == null)
-                        ? 'You need to enter the custom value as a number.'
-                        : null;
-                  },
-                ),
-              ),
-            ),
-          ],
-          //   ),
-          // ],
-        ),
-      ),
-      actionsAlignment: MainAxisAlignment.spaceEvenly,
-      actions: <Widget>[
-        OutlinedButton(
-          onPressed: onCancel,
-          child: const Text('Cancel'),
-        ),
-        OutlinedButton(
-          onPressed: onUnset,
-          child: const Text('Unset'),
-        ),
-        ActerPrimaryActionButton(
-          onPressed: onSubmit,
-          child: const Text('Submit'),
-        ),
-      ],
-    );
-  }
-
-  void onCancel() {
-    Navigator.pop(context, widget.currentPowerLevel);
-  }
-
-  void onUnset() {
-    Navigator.pop(context, null);
-  }
-
-  void onSubmit() {
-    if (!_formKey.currentState!.validate()) return;
-    final freshMemberStatus = widget.currentPowerLevelName;
-    if (freshMemberStatus == currentMemberStatus) {
-      // nothing to do, all the same.
-      Navigator.pop(context, null);
-      return;
-    }
-    int? newValue;
-    if (currentMemberStatus == 'Admin') {
-      newValue = 100;
-    } else if (currentMemberStatus == 'Mod') {
-      newValue = 50;
-    } else if (currentMemberStatus == 'Regular') {
-      newValue = 0;
-    } else {
-      newValue = customValue ?? 0;
-    }
-
-    if (widget.currentPowerLevel == newValue) {
-      // nothing to be done.
-      newValue = null;
-    }
-
-    Navigator.pop(context, newValue);
   }
 }
